// src/types/Weather.ts
/*
 * Copyright © 2022 Sinclair Broadcast Group
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * 
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.

 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

interface HourlyForecast {
  dateTime: number;
  dayOfWeek: string;
  heatIndex: number;
  humidity: number;
  precipitation: number;
  temp: number;
  title: string;
  ultraviolet: string;
  wind: number;
  windDirection: string;
  iconCode?: number;
  iconName?: string;
}
interface Day {
  dateTime: number;
  hiTemp: number;
  hourly: Array<HourlyForecast>;
  humidity: number;
  lowTemp: number;
  precipitation: number;
  title: string;
  ultraviolet: string;
  wind: number;
  windDirection: string;
  dayIconCode?: string;
  dayIconName?: string;
  nightIconCode?: string;
  nightIconName?: string;
}

interface LatestWeathercast {
  videoUrl: string;
}

interface SevenDayForecast {
  days: Day[];
  headerIconCode: number;
  headerTemp: number;
  headerTitle: string;
  hourlyForecasts: Array<HourlyForecast> | null;
  location: string;
}

interface CustomDay extends Day {
  headerTemp: number;
  headerTitle: string;
}

export type { HourlyForecast, Day, LatestWeathercast, SevenDayForecast, CustomDay };